# openApi.json
{"openapi":"3.1.0","info":{"title":"Todo Api","summary":"An example Todo application","description":"# Todo Api\n\nA simple todo api written in rust.\n","version":""},"paths":{"/todo/":{"get":{"description":"list all todos","responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Todo"}}}}}}},"post":{"description":"store a todo","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TodoInput"}}},"required":true},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Todo"}}}}}}},"/todo/{id}":{"get":{"description":"show a todo from the id","responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Todo"}}}}}},"put":{"description":"update a todo from the id","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TodoInput"}}},"required":true},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Todo"}}}}}},"delete":{"description":"Delete a todo from the id","responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RowsAffected"}}}}}}},"/docs/":{"get":{"description":"This documentation page.","responses":{"200":{"description":"HTML content","content":{"text/html":{"schema":{"type":"string"}}}}},"security":[{"ApiKey":[]}]}}},"components":{"schemas":{"RowsAffected":{"type":"object","required":["rows_affected"],"properties":{"rows_affected":{"type":"integer","format":"uint64","minimum":0.0}}},"Todo":{"type":"object","required":["completed","description","id","title"],"properties":{"completed":{"type":"boolean"},"description":{"type":"string"},"id":{"type":"string","format":"uuid"},"title":{"type":"string"}}},"TodoInput":{"type":"object","required":["description","title"],"properties":{"completed":{"type":["boolean","null"]},"description":{"type":"string"},"title":{"type":"string"}}}}}}